Wine Classifier
We often "google" reviews while buying or testing various products. So here we are keeping in mind the sentiments of our user and making their life a bit easier by using a classifier to get the perfect review for Wine. Assume we have a collection of wine reviews. We'd want to anticipate if a review is in favour or not for each one. This is the same as stating we'd like to create a model that takes a review and predicts whether a particular wine will be good or bad. (To put it another way, we'd like to create a model that takes a document d and predicts a class.)